Adds the monster overrides
Adds monster overrides for every hostile vanilla monster currently in
the game, and a few non-hostiles that are an adequate challenge to hunt
down.

Values are all a bit arbitary and likely to need tweaks in the long run,
and tend to operate on the assumption that a basic zed is worth 2 EXP,
with anything that's worth experience but easier than that having only 1
EXP, and a few really weak hostiles only granting 1 EXP half the time.

Monsters with the `QUEEN` flag tend to hover in the vicinity of 100 EXP,
plus or minus 20-30 in some cases for the less challenging ones.

Some monsterdrop itemgroups simply have EXP injected into them outright,
generally if they're used by only a single monster, or by a handful of
monsters close enough in terms of difficulty. Even then this could've
potentially been done in a few more cases but both approaches work.

And as a bonus, fixed some redudant descriptions on certain XP Trait
items.